feat: dockerize gnark, refactor, add url support to server
