+++ gnark-whir/README.md
@@ -1,17 +1,140 @@
 # WHIR proof verifier
 
-A command-line application performing a recursive verification of a zero-knowledge proof using [WHIR](https://eprint.iacr.org/2024/1586.pdf).
+A command-line and HTTP server application performing a recursive verification of a zero-knowledge proof using [WHIR](https://eprint.iacr.org/2024/1586.pdf).
 
 ## Usage
 
+### Command Line Interface
+
 ```bash
-go run . [flags]
+go run cmd/cli/main.go [flags]
 ```
 
-### Flags
+#### Flags
 
 - `--config` Path to the JSON configuration file containing verifier circuit parameters (default: `../noir-examples/poseidon-rounds/params_for_recursive_verifier`)
 - `--r1cs` Path to the R1CS JSON file describing the constraint system of the inner circuit (default: `../noir-examples/poseidon-rounds/r1cs.json`)
 - `--ccs` Optional path to store the constraint system object of the verifier circuit (default: empty, don't serialize)
 - `--pk` Optional path to load the Proving Key (PK) that will be used to generate proof for the verifier circuit. If not provided, PK will be generated unsafely (default: empty, generate own key)
 - `--vk` Optional path to load the Verifying Key (VK) that will be used to prove the verifier circuit. If not provided, VK will be generated unsafely (default: empty, generate own key)
+
+### HTTP Server
+
+Start the HTTP server:
+
+```bash
+go run cmd/server/main.go
+```
+
+## HTTP API
+
+This repository provides an HTTP API for verifying proofs.
+
+### Base URL
+
+```
+http://localhost:3000 (when running locally)
+```
+
+### Endpoints
+
+#### Health Check
+
+**GET** `/api/v1/ping`
+
+Returns a simple health check response.
+
+**Response:**
+```
+pong
+```
+
+#### Generic Proof Verification
+
+**POST** `/api/v1/verify`
+
+Generic endpoint for verifying proofs with trusted setup keys vk/pk generated by server.
+
+**Request:**
+Send a `multipart/form-data` POST request with the following fields:
+
+- `config` (required): JSON configuration file containing verifier circuit parameters
+- `r1cs` (required): R1CS JSON file describing the constraint system of the inner circuit
+- `pk_url` (optional): Publicly exposed url for downloading proving key
+- `vk_url` (optional): Publicly exposed url for downloading verifying key
+
+**Response:**
+- **Success (200)**: `Verification successful`
+- **Error (400)**: Error message describing the failure
+
+## Endpoints for Internal Testing (Will be removed cache implementation)
+
+We store vk/pk keys for common circuits for testing for quicker verification.
+
+#### Basic2 Circuit Verification
+
+**POST** `/api/v1/verifybasic2`
+
+Pre-configured endpoint for verifying `noir-examples/basic-2` circuit proofs using stored keys.
+
+**Request:**
+Send a `multipart/form-data` POST request with the following fields:
+
+- `config` (required): JSON configuration file containing verifier circuit parameters
+- `r1cs` (required): R1CS JSON file describing the constraint system of the inner circuit
+
+**Response:**
+- **Success (200)**: `Verification successful`
+- **Error (400)**: Error message describing the failure
+
+#### Age Check Circuit Verification
+
+**POST** `/api/v1/verifyagecheck`
+
+Pre-configured endpoint for verifying Age Check circuit proofs using stored vk/pk.
+
+**Request:**
+Send a `multipart/form-data` POST request with the following fields:
+
+- `config` (required): JSON configuration file containing verifier circuit parameters
+- `r1cs` (required): R1CS JSON file describing the constraint system of the inner circuit
+
+**Response:**
+- **Success (200)**: `Verification successful`
+- **Error (400)**: Error message describing the failure
+
+### Server Configuration
+
+The server is configured with the following settings:
+
+- **Port**: 3000
+- **Read Timeout**: 10 minutes (for file uploads)
+- **Write Timeout**: 5 minutes (for responses)
+- **Idle Timeout**: 90 minutes (total connection time)
+- **Body Limit**: 2GB (total size for params and R1CS files)
+- **CORS**: Enabled with permissive settings
+
+### Example Usage
+
+#### Using `curl` for Generic Verification
+
+```bash
+curl -X POST \
+  -F "config=@params_for_recursive_verifier" \
+  -F "r1cs=@r1cs.json" \
+  http://localhost:3000/api/v1/verify
+```
+
+### Error Handling
+
+The API returns appropriate HTTP status codes:
+
+- **200**: Verification successful
+- **400**: Bad request (missing files, verification failed, etc.)
+- **500**: Internal server error
+
+### File Requirements
+
+- **config**: Must be a valid JSON file containing verifier circuit parameters
+- **r1cs**: Must be a valid R1CS JSON file describing the constraint system
+
