+++ gnark-whir/app/circuit/matrix_evaluation.go
@@ -1,4 +1,4 @@
-package main
+package circuit
 
 import (
 	"math/big"
