+++ gnark-whir/app/circuit/whir_utilities.go
@@ -1,0 +1,189 @@
+package circuit
+
+import (
+	"math/bits"
+	"reilabs/whir-verifier-circuit/app/utilities"
+
+	"github.com/consensys/gnark/frontend"
+	"github.com/consensys/gnark/std/math/uints"
+	gnarkNimue "github.com/reilabs/gnark-nimue"
+	skyscraper "github.com/reilabs/gnark-skyscraper"
+)
+
+func verifyMerkleTreeProofs(api frontend.API, uapi *uints.BinaryField[uints.U64], sc *skyscraper.Skyscraper, leafIndexes []uints.U64, leaves [][]frontend.Variable, leafSiblingHashes []frontend.Variable, authPaths [][]frontend.Variable, rootHash frontend.Variable) error {
+	numOfLeavesProved := len(leaves)
+	for i := range numOfLeavesProved {
+		treeHeight := len(authPaths[i]) + 1
+		leafIndexBits := api.ToBinary(uapi.ToValue(leafIndexes[i]), treeHeight)
+		leafSiblingHash := leafSiblingHashes[i]
+
+		claimedLeafHash := sc.CompressV2(leaves[i][0], leaves[i][1])
+		for x := range len(leaves[i]) - 2 {
+			claimedLeafHash = sc.CompressV2(claimedLeafHash, leaves[i][x+2])
+		}
+
+		dir := leafIndexBits[0]
+
+		xLeftChild := api.Select(dir, leafSiblingHash, claimedLeafHash)
+		xRightChild := api.Select(dir, claimedLeafHash, leafSiblingHash)
+
+		currentHash := sc.CompressV2(xLeftChild, xRightChild)
+
+		for level := 1; level < treeHeight; level++ {
+			indexBit := leafIndexBits[level]
+
+			siblingHash := authPaths[i][level-1]
+
+			dir := api.And(indexBit, 1)
+			left := api.Select(dir, siblingHash, currentHash)
+			right := api.Select(dir, currentHash, siblingHash)
+
+			currentHash = sc.CompressV2(left, right)
+		}
+		api.AssertIsEqual(currentHash, rootHash)
+	}
+	return nil
+}
+
+func getStirChallenges(
+	api frontend.API,
+	arthur gnarkNimue.Arthur,
+	numQueries int,
+	domainSize int,
+	foldingFactorPower int,
+) ([]frontend.Variable, error) {
+	foldedDomainSize := domainSize / foldingFactorPower
+	domainSizeBytes := (bits.Len(uint(foldedDomainSize*2-1)) - 1 + 7) / 8
+
+	stirQueries := make([]uints.U8, domainSizeBytes*numQueries)
+	if err := arthur.FillChallengeBytes(stirQueries); err != nil {
+		return nil, err
+	}
+
+	bitLength := bits.Len(uint(foldedDomainSize)) - 1
+
+	indexes := make([]frontend.Variable, numQueries)
+	for i := range numQueries {
+		var value frontend.Variable = 0
+		for j := range domainSizeBytes {
+			value = api.Add(stirQueries[j+i*domainSizeBytes].Val, api.Mul(value, 256))
+		}
+
+		bitsOfValue := api.ToBinary(value)
+		indexes[i] = api.FromBinary(bitsOfValue[:bitLength]...)
+	}
+
+	return indexes, nil
+}
+
+func generateEmptyMainRoundData(circuit WHIRParams) MainRoundData {
+	return MainRoundData{
+		OODPoints:             make([][]frontend.Variable, len(circuit.RoundParametersOODSamples)),
+		StirChallengesPoints:  make([][]frontend.Variable, len(circuit.RoundParametersOODSamples)),
+		CombinationRandomness: make([][]frontend.Variable, len(circuit.RoundParametersOODSamples)),
+	}
+}
+
+func fillInOODPointsAndAnswers(numberOfOODPoints int, arthur gnarkNimue.Arthur) ([]frontend.Variable, []frontend.Variable, error) {
+	oodPoints := make([]frontend.Variable, numberOfOODPoints)
+	oodAnswers := make([]frontend.Variable, numberOfOODPoints)
+
+	if err := arthur.FillChallengeScalars(oodPoints); err != nil {
+		return nil, nil, err
+	}
+
+	if err := arthur.FillNextScalars(oodAnswers); err != nil {
+		return nil, nil, err
+	}
+
+	return oodPoints, oodAnswers, nil
+}
+
+func runWhirSumcheckRounds(
+	api frontend.API,
+	lastEval frontend.Variable,
+	arthur gnarkNimue.Arthur,
+	foldingFactor int,
+	polynomialDegree int,
+) ([]frontend.Variable, frontend.Variable, error) {
+	sumcheckPolynomial := make([]frontend.Variable, polynomialDegree)
+	foldingRandomness := make([]frontend.Variable, foldingFactor)
+	foldingRandomnessTemp := make([]frontend.Variable, 1)
+
+	for i := range foldingFactor {
+		if err := arthur.FillNextScalars(sumcheckPolynomial); err != nil {
+			return nil, nil, err
+		}
+		if err := arthur.FillChallengeScalars(foldingRandomnessTemp); err != nil {
+			return nil, nil, err
+		}
+		foldingRandomness[i] = foldingRandomnessTemp[0]
+		utilities.CheckSumOverBool(api, lastEval, sumcheckPolynomial)
+		lastEval = utilities.EvaluateQuadraticPolynomialFromEvaluationList(api, sumcheckPolynomial, foldingRandomness[i])
+	}
+	return foldingRandomness, lastEval, nil
+}
+
+func computeWPoly(
+	api frontend.API,
+	circuit WHIRParams,
+	initialData InitialSumcheckData,
+	mainRoundData MainRoundData,
+	totalFoldingRandomness []frontend.Variable,
+	linearStatementValuesAtPoints []frontend.Variable,
+) frontend.Variable {
+	numberVars := circuit.MVParamsNumberOfVariables
+
+	value := frontend.Variable(0)
+	for j := range initialData.InitialOODQueries {
+		value = api.Add(value, api.Mul(initialData.InitialCombinationRandomness[j], utilities.EqPolyOutside(api, utilities.ExpandFromUnivariate(api, initialData.InitialOODQueries[j], numberVars), totalFoldingRandomness)))
+	}
+
+	for j, linearStatementValueAtPoint := range linearStatementValuesAtPoints {
+		value = api.Add(value, api.Mul(initialData.InitialCombinationRandomness[len(initialData.InitialOODQueries)+j], linearStatementValueAtPoint))
+	}
+	for r := range mainRoundData.OODPoints {
+		numberVars -= circuit.FoldingFactorArray[r]
+		newTmpArr := append(mainRoundData.OODPoints[r], mainRoundData.StirChallengesPoints[r]...)
+
+		sumOfClaims := frontend.Variable(0)
+		for i := range newTmpArr {
+			point := utilities.ExpandFromUnivariate(api, newTmpArr[i], numberVars)
+			sumOfClaims = api.Add(sumOfClaims, api.Mul(utilities.EqPolyOutside(api, point, totalFoldingRandomness[0:numberVars]), mainRoundData.CombinationRandomness[r][i]))
+		}
+		value = api.Add(value, sumOfClaims)
+	}
+
+	return value
+}
+
+func fillInAndVerifyRootHash(
+	roundNum int,
+	api frontend.API,
+	uapi *uints.BinaryField[uints.U64],
+	sc *skyscraper.Skyscraper,
+	circuit Merkle,
+	arthur gnarkNimue.Arthur,
+) error {
+	rootHash := make([]frontend.Variable, 1)
+	if err := arthur.FillNextScalars(rootHash); err != nil {
+		return err
+	}
+	err := verifyMerkleTreeProofs(api, uapi, sc, circuit.LeafIndexes[roundNum], circuit.Leaves[roundNum], circuit.LeafSiblingHashes[roundNum], circuit.AuthPaths[roundNum], rootHash[0])
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func computeFold(leaves [][]frontend.Variable, foldingRandomness []frontend.Variable, api frontend.API) []frontend.Variable {
+	computedFold := make([]frontend.Variable, len(leaves))
+	for j := range leaves {
+		computedFold[j] = utilities.MultivarPoly(leaves[j], foldingRandomness, api)
+	}
+	return computedFold
+}
+
+func calculateShiftValue(oodAnswers []frontend.Variable, combinationRandomness []frontend.Variable, computedFold []frontend.Variable, api frontend.API) frontend.Variable {
+	return utilities.DotProduct(api, append(oodAnswers, computedFold...), combinationRandomness)
+}
