+++ gnark-whir/docker-compose.yml
@@ -1,0 +1,41 @@
+services:
+  verifier:
+    build:
+      context: .
+      dockerfile: Dockerfile
+      args:
+        TARGETOS: linux
+        TARGETARCH: arm64
+    container_name: gnark-whir-verifier
+    ports:
+      - "3000:3000"
+    deploy:
+      resources:
+        limits:
+          memory: 20G
+        reservations:
+          memory: 10G
+    memswap_limit: 20G
+    shm_size: 3g
+    ulimits:
+      nofile:
+        soft: 65536
+        hard: 65536
+      nproc:
+        soft: 32768
+        hard: 32768
+    restart: unless-stopped
+    healthcheck:
+      test: ["CMD", "wget", "--no-verbose", "--tries=1", "--spider", "http://localhost:3000/api/v1/ping"]
+      interval: 30s
+      timeout: 3s
+      retries: 3
+      start_period: 5s
+    volumes:
+      - ./keys:/app/keys
+    networks:
+      - gnark-network
+
+networks:
+  gnark-network:
+    driver: bridge
